Add files via upload
Text Editor built using Java uses JTextArea, JMenu, JMenuItems, and JMenuBar to perform various tasks. It allows the user to enter, change, store, and print text. It also has a file menu to make changes in files (like open, save, close, and print) for future reference. and an edit menu to cut, copy and paste texts. Also, it has a “Save and Submit” button to close the file after saving the data. An actionListener is also used to detect actions in the project.